requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-multiselectfield==0.1.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
